requirements.txt
flask==3.0.3
jinja2==3.1.4
werkzeug==3.0.3
cachetools
beautifulsoup4
pillow
imagehash
hachoir
requests
gunicorn